card-tracker-scraper/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.6
supabase==2.15.0
google-generativeai==0.8.5
